refactor: schema 폴더 구조화 (#23)
