+++ src/app/admin/(with-sidebar)/results/_components/result-manage/ResultTable.tsx
@@ -1,5 +1,5 @@
 import {RESULT_PARTS} from '@/constants/admin/admin-result';
-import {ResultSummaryData} from '@/schemas/admin-result-type';
+import {ResultSummaryData} from '@/schemas/admin/admin-result-type';
 
 interface ResultTableProps {
   data: ResultSummaryData[];
