--- src/schemas/admin-application-type.ts
@@ -1,5 +1,0 @@
-export type PartType = 'all' | 'plan' | 'design' | 'frontend' | 'backend';
-
-export type PartTypeEtc = 'etc' | 'plan' | 'design' | 'frontend' | 'backend';
-
-export type ReviewerType = 'admin1' | 'admin2' | 'admin3' | 'admin4';
