+++ src/schemas/admin/admin-result-type.ts
